step2/src/loadgen/pyproject.toml
[tool.poetry]
name = "loadgen"
version = "0.1.0"
description = "Load generator for Shakesapp client"
authors = ["Yoshi Yamaguchi <yoshifumi@google.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
structlog = "^21.1.0"
requests = "^2.25.1"
opentelemetry-propagator-gcp = "^1.0.0rc0"
opentelemetry-exporter-gcp-trace = "^1.0.0rc0"
opentelemetry-instrumentation-requests = "0.20b0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
